requirements.txt
apscheduler==3.11.1
dbutils==3.1.2
fastapi==0.127.0
pydantic==2.12.5
PyJWT==2.10.1
Requests==2.32.5
slowapi==0.1.9
suds-py3==1.4.5
uvicorn==0.40.0
python-multipart
requests==2.32.5
sqlalchemy==2.0.27
